+++ mesh.h
@@ -21,7 +21,7 @@ class Mesh
     int getCellCount() const;
 
     std::vector<Cell*> getCellsWithVertex( const std::string& ) const;
-    std::vector<Cell*> getCellsWithVertex( const int, const int ) const;
+    std::vector<Cell*> getCellsWithVertex( const double, const double ) const;
 
     friend std::ostream& operator<<( std::ostream&, const Mesh& );
 };
